uint16

+++ ziface/iconnection.go
@@ -29,10 +29,10 @@ type IConnection interface {
 	GetConnID() uint32                 //获取当前连接ID
 	RemoteAddr() net.Addr              //获取远程客户端地址信息
 
-	SendMsg(msgID uint32, msgType int, data []byte) error     //直接将Message数据发送数据给远程的TCP客户端(无缓冲)
-	SendBuffMsg(msgID uint32, msgType int, data []byte) error //直接将Message数据发送给远程的TCP客户端(有缓冲)
-	SendBinaryMsg(msgID uint32, data []byte) error
-	SendBinaryBuffMsg(msgID uint32, data []byte) error
+	SendMsg(msgID uint16, msgType int, data []byte) error     //直接将Message数据发送数据给远程的TCP客户端(无缓冲)
+	SendBuffMsg(msgID uint16, msgType int, data []byte) error //直接将Message数据发送给远程的TCP客户端(有缓冲)
+	SendBinaryMsg(msgID uint16, data []byte) error
+	SendBinaryBuffMsg(msgID uint16, data []byte) error
 
 	SetProperty(key string, value interface{})   //设置链接属性
 	GetProperty(key string) (interface{}, error) //获取链接属性
